+++ requirements/base.txt
@@ -21,10 +21,8 @@ django-extensions==3.1.3
 django-model-utils==4.3.1
 
 # Time zones support - Do not update these without doing significant testing!
-# pytz==2018.7
-# python-dateutil==2.7.3
-pytz==2024.1
-python-dateutil==2.9.0.post0
+pytz==2018.7
+python-dateutil==2.7.3
 
 # Metric/imperial units support
 django-pint==0.6
@@ -68,8 +66,7 @@ seed-salesforce==0.1.0
 # geospatial and pnnl/buildingid-py
 shapely==2.0.1
 usaddress==0.5.10
-# -e git+https://github.com/SEED-platform/buildingid.git@603300c8a7212b339b7e8cb7a4365be6726195aa#egg=pnnl-buildingid
--e git+https://github.com/SEED-platform/buildingid.git@8d59ec016921d47e998a78339a8195913ee36bee#egg=pnnl-buildingid
+-e git+https://github.com/SEED-platform/buildingid.git@603300c8a7212b339b7e8cb7a4365be6726195aa#egg=pnnl-buildingid
 
 
 oauthlib==2.0.3
